Tools: autotest: move close method up to common.py
