+++ Tools/autotest/arduplane.py
@@ -4,7 +4,6 @@
 from __future__ import print_function
 import math
 import os
-import shutil
 
 import pexpect
 from pymavlink import mavutil
@@ -47,6 +46,8 @@ class AutoTestPlane(AutoTest):
 
         self.sitl = None
         self.hasInit = False
+
+        self.log_name = "ArduPlane"
 
     def init(self):
         if self.frame is None:
@@ -100,22 +101,6 @@ class AutoTestPlane(AutoTest):
         self.mav.idle_hooks.append(self.idle_hook)
         self.hasInit = True
         self.progress("Ready to start testing!")
-
-    def close(self):
-        if self.use_map:
-            self.mavproxy.send("module unload map\n")
-            self.mavproxy.expect("Unloaded module map")
-
-        self.mav.close()
-        util.pexpect_close(self.mavproxy)
-        util.pexpect_close(self.sitl)
-
-        valgrind_log = util.valgrind_log_filepath(binary=self.binary,
-                                                  model=self.frame)
-        if os.path.exists(valgrind_log):
-            os.chmod(valgrind_log, 0o644)
-            shutil.copy(valgrind_log,
-                        self.buildlogs_path("ArduPlane-valgrind.log"))
 
     def takeoff(self):
         """Takeoff get to 30m altitude."""
