+++ Tools/autotest/ardusub.py
@@ -3,7 +3,6 @@
 # Dive ArduSub in SITL
 from __future__ import print_function
 import os
-import shutil
 
 import pexpect
 from pymavlink import mavutil
@@ -46,6 +45,8 @@ class AutoTestSub(AutoTest):
 
         self.sitl = None
         self.hasInit = False
+
+        self.log_name = "ArduSub"
 
     def init(self):
         if self.frame is None:
@@ -123,22 +124,6 @@ class AutoTestSub(AutoTest):
         self.mav.idle_hooks.append(self.idle_hook)
         self.hasInit = True
         self.progress("Ready to start testing!")
-
-    def close(self):
-        if self.use_map:
-            self.mavproxy.send("module unload map\n")
-            self.mavproxy.expect("Unloaded module map")
-
-        self.mav.close()
-        util.pexpect_close(self.mavproxy)
-        util.pexpect_close(self.sitl)
-
-        valgrind_log = util.valgrind_log_filepath(binary=self.binary,
-                                                  model=self.frame)
-        if os.path.exists(valgrind_log):
-            os.chmod(valgrind_log, 0o644)
-            shutil.copy(valgrind_log,
-                        self.buildlogs_path("ArduSub-valgrind.log"))
 
     def dive_manual(self):
         self.set_rc(3, 1600)
